+++ src/module/event/listener/index.ts
@@ -1,1 +1,1 @@
-export * as listener from './event';
+export * as listener from './listener';
